+++ libs/docs/platform/standard-list-item/examples/platform-standard-list-item-border-less-example.component.ts
@@ -32,7 +32,7 @@ export class PlatformStandardListItemBorderLessExampleComponent {
                 'is a battery- or AC-powered personal computer generally smaller than a briefcase that can ' +
                 'easily be transported and conveniently',
             avatar: {
-                image: 'https://picsum.photos/400/400?nature',
+                image: 'https://picsum.photos/seed/nature/400/400',
                 ariaLabel: 'nature'
             }
         },
@@ -51,14 +51,14 @@ export class PlatformStandardListItemBorderLessExampleComponent {
             description: 'First text item in Byline (Standard text item)',
             secondary: 'Second text item in Byline (Can be semantic (Status) or not)',
             avatar: {
-                image: 'https://picsum.photos/400/400?nature',
+                image: 'https://picsum.photos/seed/nature/400/400',
                 ariaLabel: 'nature'
             }
         },
         {
             title: 'Item4',
             avatar: {
-                image: 'https://picsum.photos/400/400?nature',
+                image: 'https://picsum.photos/seed/nature/400/400',
                 ariaLabel: 'nature'
             },
             description: 'First text item in Byline (Standard text item)',